fix(testutil): not split testutil into a separate module
